models/requirement.txt
numpy
tensorflow-serving-api==2.11.0
pyvips==2.2.1
openvino==2022.3
